+++ src/didery/static/transcrypt/components/tabs.py
@@ -1,45 +1,35 @@
 # ================================================== #
-#                      ROUTER                        #
+#                        TABS                        #
 # ================================================== #
 # Author: Brady Hammond                              #
-# Created: 04/26/2018                                #
+# Created: 05/08/2018                                #
 # Last Edited:                                       #
 # Last Edited By:                                    #
 # ================================================== #
 #                      IMPORTS                       #
 # ================================================== #
 
-import dashboard
+import components.tabledtab as tabledtab
+import components.tables as tables
 
 # ================================================== #
 #                  CLASS DEFINITIONS                 #
 # ================================================== #
 
-class Router:
+class Errors(tabledtab.TabledTab):
     """
-    Class for routing between urls and pages.
+    Class for errors tab.
     """
-    def __init__(self):
-        """
-        Initialize Router object. Load navigation tabs.
-        """
-        self.tabs = dashboard.Manager()
+    Name = "Errors"
+    Icon = "i.exclamation.circle.icon"
+    DataTab = "errors"
+    Active = True
 
-    # ============================================== #
-
-    def route(self, root=None):
+    def setup_table(self):
         """
-        Sets up project routes.
-
-            Parameters:
-            root - DOM for root page
+        Sets up errors table.
         """
-        if root is None:
-            root = document.body
-        m.route(root, "/dashboard",
-            {
-                "/dashboard": {"render": self.tabs.view}
-            })
+        self.table = tables.ErrorsTable()
 
 # ================================================== #
 #                        EOF                         #
